Restructure configuration and constants in a module
Avoid importing cloud related configuration in
cull expired cluster service.
