+++ docker-compose.yml
@@ -38,10 +38,6 @@ services:
         source: ./run/configuration.json
         target: /home/mcu/configuration.json
         read_only: true
-      - type: bind
-        source: ./run/clouds.yaml
-        target: /home/mcu/credentials/clouds.yaml
-        read_only: true
 
 volumes:
   database:
