+++ mchub/models/magic_castle/magic_castle.py
@@ -19,11 +19,12 @@ from .. terraform.terraform_state_parser import TerraformStateParser
 from .. terraform.terraform_plan_parser import TerraformPlanParser
 from .. cloud.dns_manager import DnsManager
 from .. puppet.provisioning_manager import ProvisioningManager, MAX_PROVISIONING_TIME
-from ... constants import (
+
+from ... configuration.magic_castle import (
     MAIN_TERRAFORM_FILENAME,
     TERRAFORM_STATE_FILENAME,
-    CLUSTERS_PATH
 )
+from ... configuration.env import CLUSTERS_PATH
 
 from ... exceptions.invalid_usage_exception import *
 from ... exceptions.server_exception import *
