+++ mchub/models/user/authenticated_user.py
@@ -3,7 +3,7 @@ import json
 from . user import User
 from .. magic_castle.magic_castle import MagicCastle
 from ... configuration import config
-from ... constants import DEFAULT_CLOUD
+from ... configuration.cloud import DEFAULT_CLOUD
 
 
 class AuthenticatedUser(User):
