+++ tests/integration/test_api_no_auth.py
@@ -1,4 +1,4 @@
-from mchub.constants import DEFAULT_CLOUD
+from mchub.configuration.cloud import DEFAULT_CLOUD
 from mchub.models.magic_castle.cluster_status_code import ClusterStatusCode
 from mchub import create_app
